refactor code to be pip installable (under single versionable package)
