+++ hmgmetrics/__init__.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+"""
+hmgmetrics
+------------
+
+Human Motion Generation Metrics
+Unifyied framework for human motion generation evaluation metrics. _[1]
+
+[1]: A. Ismail-Fawaz, M. Devanne, S. Berretti, J. Weber, and G. Forestier,
+ “Establishing a Unified Evaluation Framework for Human Motion Generation:
+ A Comparative Analysis of Metrics,” May 13, 2024, arXiv: arXiv:2405.07680.
+ Accessed: May 16, 2024. [Online]. Available: http://arxiv.org/abs/2405.07680
+
+
+:copyright: (c) 2024 by Cognitive Systems Lab.
+:license: MIT
+"""
+# Imports
+
+# built-in
+
+# local
+from .__version__ import __version__  # noqa: F401
+
+# 3rd-party
