+++ hmgmetrics/__version__.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8 -*-
+"""
+metrics.__version__
+-------------------
+
+Version tracker.
+
+:copyright: (c) 2024 by Cognitive Systems Lab.
+:license: MIT
+"""
+# Imports
+
+# built-in
+
+# local
+
+# 3rd-party
+
+# CSL
+
+
+__VERSION__ = (1, 0, 0)
+
+__version__ = '.'.join(map(str, __VERSION__))
