--- metrics/diversity/__init__.py
@@ -1,5 +1,0 @@
-from metrics.diversity._apd import APD
-from metrics.diversity._acpd import ACPD
-from metrics.diversity._coverage import COVERAGE
-from metrics.diversity._wpd import WPD
-from metrics.diversity._mms import MMS
